# performance improvements when rendering tables